requirement.txt
absl-py==0.7.1
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
audioread==2.1.8
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bandmat==0.7
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
boto3==1.9.156
botocore==1.12.156
Bottleneck==1.2.1
bz2file==0.98
cachetools==3.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.3.9
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
contextlib2==0.5.5
convertdate==2.1.3
cryptography==2.2.2
cycler==0.10.0
cysignals==1.10.2
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
dill==0.3.0
distributed==1.21.8
docutils==0.14
dopamine-rl==2.0.5
entrypoints==0.2.3
enum34==1.1.6
ephem==3.7.6.0
et-xmlfile==1.0.1
fastcache==1.0.2
fbprophet==0.5
ffmpeg==1.4
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
future==0.17.1
gast==0.2.2
gensim==3.7.3
gevent==1.3.0
gin-config==0.1.4
glob2==0.6
gmpy2==2.0.8
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.6.0
greenlet==0.4.13
grpcio==1.20.1
gunicorn==19.9.0
gym==0.13.0
h5py==2.9.0
heapdict==1.0.0
holidays==0.9.10
html5lib==1.0.1
httplib2==0.13.0
hyperopt==0.1.2
idna==2.6
imageio==2.3.0
imagesize==1.0.0
imbalanced-learn==0.4.3
imblearn==0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kfac==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.3
llvmlite==0.23.1
locket==0.2.0
lunardate==0.2.0
lxml==4.2.1
Markdown==3.1
MarkupSafe==1.0
MatchZoo==2.1.0
matplotlib==2.2.2
mccabe==0.6.1
mesh-tensorflow==0.0.5
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
oauth2client==4.1.3
odo==0.5.1
olefile==0.45.1
opencv-contrib-python-headless==4.1.0.25
opencv-python==4.1.0.25
openpyxl==2.5.3
packaging==17.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
promise==2.2.1
prompt-toolkit==1.0.15
protobuf==3.7.1
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycocotools==2.0.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydub==0.23.1
pyfasttext==0.4.6
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.8.4
pymongo==3.8.0
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypng==0.0.19
PySocks==1.6.8
pystan==2.18.0.0
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
pytest-runner==5.1
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.18.4
resampy==0.2.1
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
s3transfer==0.2.0
scikit-image==0.13.1
scikit-learn==0.21.2
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
setuptools-git==1.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.8.3
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensor2tensor==1.13.4
tensorboard==1.12.2
tensorflow-datasets==1.0.2
tensorflow-gpu==1.12.0
tensorflow-metadata==0.13.0
tensorflow-probability==0.5.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
toolz==0.9.0
torch==1.1.0
torchvision==0.2.2.post3
tornado==5.0.2
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xgboost==0.90
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zict==0.1.3
